Dispatch a download task after a sync if policy is 'background'.
closes #1445
