+++ server/pulp/plugins/conduits/mixins.py
@@ -45,13 +45,6 @@ class ContentSourcesConduitException(Exception):
     pass
 
 
-class LazyStatusConduitException(Exception):
-    """
-    General exception that wraps any exception coming out of the Pulp server.
-    """
-    pass
-
-
 class RepoScratchPadMixin(object):
 
     def __init__(self, repo_id, exception_class):
