+++ server/pulp/server/tasks/__init__.py
@@ -6,3 +6,4 @@ of this Python package.
 import pulp.server.db.reaper  # noqa
 import pulp.server.maintenance.monthly  # noqa
 import pulp.server.controllers.content  # noqa
+import pulp.server.controllers.repository  # noqa
