+++ server/test/unit/server/webservices/views/test_repositories.py
@@ -1725,9 +1725,9 @@ class TestRepoDownload(unittest.TestCase):
 
     @mock.patch('pulp.server.webservices.views.decorators._verify_auth',
                 new=assert_auth_EXECUTE())
-    @mock.patch('pulp.server.webservices.views.repositories.content_controller')
+    @mock.patch('pulp.server.webservices.views.repositories.repo_controller')
     @mock.patch('pulp.server.webservices.views.repositories.model.Repository.objects')
-    def test_post_download_repo(self, mock_repo_qs, mock_content_controller):
+    def test_post_download_repo(self, mock_repo_qs, mock_repo_controller):
         """Test that a repo download task is dispatched."""
         # Setup
         download_repo = RepoDownload()
@@ -1737,16 +1737,16 @@ class TestRepoDownload(unittest.TestCase):
             download_repo.post(mock.Mock(body=None), 'mock_repo')
         self.assertEqual(cm.exception.http_status_code, 202)
         mock_repo_qs.get_repo_or_missing_resource.assert_called_once_with('mock_repo')
-        mock_content_controller.queue_download_repo.assert_called_once_with(
+        mock_repo_controller.queue_download_repo.assert_called_once_with(
             'mock_repo',
             verify_all_units=False
         )
 
     @mock.patch('pulp.server.webservices.views.decorators._verify_auth',
                 new=assert_auth_EXECUTE())
-    @mock.patch('pulp.server.webservices.views.repositories.content_controller')
+    @mock.patch('pulp.server.webservices.views.repositories.repo_controller')
     @mock.patch('pulp.server.webservices.views.repositories.model.Repository.objects')
-    def test_post_download_repo_verify(self, mock_repo_qs, mock_content_controller):
+    def test_post_download_repo_verify(self, mock_repo_qs, mock_repo_controller):
         """Test that a repo download task is dispatched with verify_all_units."""
         # Setup
         mock_request = mock.Mock(body=json.dumps({'verify_all_units': True}))
@@ -1757,16 +1757,16 @@ class TestRepoDownload(unittest.TestCase):
             download_repo.post(mock_request, 'mock_repo')
         self.assertEqual(cm.exception.http_status_code, 202)
         mock_repo_qs.get_repo_or_missing_resource.assert_called_once_with('mock_repo')
-        mock_content_controller.queue_download_repo.assert_called_once_with(
+        mock_repo_controller.queue_download_repo.assert_called_once_with(
             'mock_repo',
             verify_all_units=True
         )
 
     @mock.patch('pulp.server.webservices.views.decorators._verify_auth',
                 new=assert_auth_EXECUTE())
-    @mock.patch('pulp.server.webservices.views.repositories.content_controller')
+    @mock.patch('pulp.server.webservices.views.repositories.repo_controller')
     @mock.patch('pulp.server.webservices.views.repositories.model.Repository.objects')
-    def test_post_download_repo_bad_request(self, mock_repo_qs, mock_content_controller):
+    def test_post_download_repo_bad_request(self, mock_repo_qs, mock_repo_controller):
         """Test that a repo download call with a bad request results in a 400."""
         # Setup
         mock_request = mock.Mock(body=json.dumps({'verify_all_units': 'please'}))
@@ -1778,13 +1778,13 @@ class TestRepoDownload(unittest.TestCase):
         self.assertEqual(error_codes.PLP1010, cm.exception.error_code)
         self.assertEqual(cm.exception.http_status_code, 400)
         mock_repo_qs.get_repo_or_missing_resource.assert_called_once_with('mock_repo')
-        self.assertEqual(0, mock_content_controller.queue_download_repo.call_count)
+        self.assertEqual(0, mock_repo_controller.queue_download_repo.call_count)
 
     @mock.patch('pulp.server.webservices.views.decorators._verify_auth',
                 new=assert_auth_EXECUTE())
-    @mock.patch('pulp.server.webservices.views.repositories.content_controller')
+    @mock.patch('pulp.server.webservices.views.repositories.repo_controller')
     @mock.patch('pulp.server.webservices.views.repositories.model.Repository.objects')
-    def test_post_download_repo_missing_repo(self, mock_repo_qs, mock_content_controller):
+    def test_post_download_repo_missing_repo(self, mock_repo_qs, mock_repo_controller):
         """Test that a repo download call with an invalid repo id results in a 404."""
         # Setup
         mock_repo_qs.get_repo_or_missing_resource.side_effect = exceptions.MissingResource
@@ -1795,7 +1795,7 @@ class TestRepoDownload(unittest.TestCase):
             download_repo.post(mock.Mock(body=None), 'mock_repo')
         self.assertEqual(cm.exception.http_status_code, 404)
         mock_repo_qs.get_repo_or_missing_resource.assert_called_once_with('mock_repo')
-        self.assertEqual(0, mock_content_controller.queue_download_repo.call_count)
+        self.assertEqual(0, mock_repo_controller.queue_download_repo.call_count)
 
 
 class TestRepoAssociate(unittest.TestCase):
